初始化游戏参数

+++ src/gameType/chuachua/entity/Brick.java
@@ -3,8 +3,8 @@ package gameType.chuachua.entity;
 public class Brick {
 	public int height;//高
 	public int width;//宽
-	public int locX;//x
-	public int locY;//y
-	public String bPropsId;//砖块道具id
+	public double locX;//x
+	public double locY;//y
+	public String bPropsId ;//砖块道具id
 	public int hardness;//硬度
 }
